# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Portuguese accommodations (Alojamentos Locais) data API",
        "title": "Pedro API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@pedroapi.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8087",
    "basePath": "/",
    "paths": {
        "/admin": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Access admin area (requires authentication)",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Admin access",
                "responses": {
                    "200": {
                        "description": "admin",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/alojamentos": {
            "get": {
                "description": "Get a paginated list of Portuguese accommodations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alojamentos"
                ],
                "summary": "List accommodations with pagination",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default: 20, max: 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field (id, nr_rnal, denominacao, concelho, distrito, created_at)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order (asc, desc)",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PaginatedResponse-models_AlojamentoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/alojamentos/search": {
            "get": {
                "description": "Search accommodations with various filters and pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alojamentos"
                ],
                "summary": "Search accommodations with filters",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default: 20, max: 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field (id, nr_rnal, denominacao, concelho, distrito, created_at)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order (asc, desc)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by municipality",
                        "name": "concelho",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by district",
                        "name": "distrito",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by accommodation type",
                        "name": "modalidade",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by owner email",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Minimum capacity",
                        "name": "min_capacity",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum capacity",
                        "name": "max_capacity",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Minimum latitude",
                        "name": "min_lat",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Maximum latitude",
                        "name": "max_lat",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Minimum longitude",
                        "name": "min_lng",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Maximum longitude",
                        "name": "max_lng",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PaginatedResponse-models_AlojamentoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/alojamentos/stats": {
            "get": {
                "description": "Get aggregated statistics about accommodations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alojamentos"
                ],
                "summary": "Get accommodation statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StatsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/alojamentos/{id}": {
            "get": {
                "description": "Get a single accommodation by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "alojamentos"
                ],
                "summary": "Get accommodation by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Accommodation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AlojamentoResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.AlojamentoResponse": {
            "type": "object",
            "properties": {
                "codigo_postal": {
                    "type": "string"
                },
                "concelho": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "data_abertura_publico": {
                    "type": "string"
                },
                "data_registo": {
                    "type": "string"
                },
                "denominacao": {
                    "type": "string"
                },
                "distrito": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "endereco": {
                    "type": "string"
                },
                "freguesia": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "latitude": {
                    "type": "number"
                },
                "localidade": {
                    "type": "string"
                },
                "longitude": {
                    "type": "number"
                },
                "modalidade": {
                    "type": "string"
                },
                "nr_rnal": {
                    "type": "integer"
                },
                "nr_utentes": {
                    "type": "integer"
                }
            }
        },
        "models.DistrictStats": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "distrito": {
                    "type": "string"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "models.MunicipalityStats": {
            "type": "object",
            "properties": {
                "concelho": {
                    "type": "string"
                },
                "count": {
                    "type": "integer"
                }
            }
        },
        "models.PaginatedResponse-models_AlojamentoResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AlojamentoResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/models.PaginationMeta"
                }
            }
        },
        "models.PaginationMeta": {
            "type": "object",
            "properties": {
                "has_more": {
                    "type": "boolean"
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "models.StatsResponse": {
            "type": "object",
            "properties": {
                "average_capacity": {
                    "type": "number"
                },
                "by_concelho": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MunicipalityStats"
                    }
                },
                "by_distrito": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.DistrictStats"
                    }
                },
                "by_modalidade": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TypeStats"
                    }
                },
                "total_accommodations": {
                    "type": "integer"
                }
            }
        },
        "models.TypeStats": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "modalidade": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}